Add base class to hold common logic for both preprocess throttle classes
